Added a mock method for POST request, added a middleware file for user who wants to use MSL as a module in their express server.
