+++ msl-client-node/package.json
@@ -10,20 +10,21 @@
 	"xhr2":"0.0.7"
 	},
   "devDependencies": {
-	"msl-server": "~1.0.11"
+	"msl-server": "~1.0.2"
 	},
   "author": [
   "Daniel Koo <danieljamin.koo@gmail.com>",
   "Bryan Robbins <bryantrobbins@gmail.com>",
   "Chien-Hung Kuo <kaidanransha@gmail.com>",
-  "Jake Sheppard <jshep409@gmail.com>"
+  "Jake Sheppard <jshep409@gmail.com>",
+  "Liusongsen Yang <ylss0110@gmail.com>"
   ],
   "main": "mockapi.js",
   "bin": {},
   "engines": {
     "node": "~0.10"
   },
-  "version": "1.0.5",
+  "version": "1.0.6",
   "readmeFilename": "README.md",
   "license": "Apache 2.0"
 }
