+++ msl-server-node/bin/msl
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../web-server');
+require('../web-server')();
